requirement.txt
numpy
pygame
sounddevice
gtts
pyttsx3
aiohttp
aiosignal==1.3.1
anyio==3.6.2
APScheduler
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4
bilibili-api-python==16.0.0
sip
PyQt5
PyQt5-tools
python-vlc